Add a shared state backend that can be updated
The data structure holding the repository data is fixed at boot time as
dynamic reconfiguring of the repository list isn't supported at this
time. However, the value entries holding the `Repo` struct is guarded
with a `RwLock<Arc<Repo>>`. This allows the provider to return immutable
references to a repo regardless whether or not the repo data is being
updated. If multiple references are being used, the repository can still
be updated and the when the number of references to the stale copy
reaches zero, then that data will be freed. A repository is updated with
an `Arc::make_mut` providing copy-on-write like behaviors if other
readers are still active (see:
https://twitter.com/sadisticsystems/status/1155817005947207680).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>