Add GitHub client & implement populate/update
This work adds GitHub HTTP and API clients which fetch the metadata
about registered repositories. The clients are asynchronous, returning
futures which allow the API to be used nicely with the updater
infrastructure, running on the tokio runtime. Somewhat unfortunetly, a
decent amount of work went into the first iteration using actix's built
in client "awc" which was a very nice experience until I realized that
the client doesn't yet support following redirects which was a hard
requirement for querying GitHub release asset bodies. Work transitioned
to using the "reqwest" client's async API. Thankfully both versions
supported and worked with rustls for TLS meaning that OpenSSL doesn't
have to be in the compilation chain.

There is a fair amount of error handling code which is yet to be plumbed
in, meaning a lot of `.expect("TODO: ...")` and `panic!("TODO: ...")`
calls still exist. This is will addressed in future work.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>