Pull up update logging & reduce allocs/copies in futures
This change started with pulling up the logging for updates to `Updater`
which mainly involved pushing errors further upstack. This resulted in
better error logging when GitHub repositories were not found on updates
and it also exposed a seam which highlighted some now-unnecessary `Arc`s
and String clones.

With a bit of thought and some searching through `.clone()` calls, it
became possible to allocate far fewer String clones for the underlying
futures (remember, they need ownership or `&'static` lifetime borrows)
by introducing a couple `Arc<str>`s, knowing that the code only needed
these string references for querying or reporting--not to mutate.
Cloning `Arc<str>` is cheaper memory-wise and there should be less
chance of accidentally mutating a string deeper in code needlessly.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>