Add custom deserialization for `Config`
This implementation uses a custom deserializer to determine the provider
type and then deserialize to the correct provider variant. Currently the
`bind_addr` and `oauth_token` fields support environment variable
interpolation, meaning that one or more shell-style variables can be
present in the value string and they will be replaced with the
environment variable value. The ultimate configuration will be read in
as a JSON file but that has yet to be implemented. In the meantime a
stubbed out JSON string is constructed and passed into the
`app::config()` function.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>